fixed problem with detecting video error
