+++ App/Mutuo.Etl/Blob/AzureBlobFileStore.cs
@@ -10,6 +10,7 @@ using Humanizer;
 using Microsoft.Azure.Storage;
 using Microsoft.Azure.Storage.Blob;
 using Newtonsoft.Json;
+using Serilog;
 using SysExtensions;
 using SysExtensions.Collections;
 using SysExtensions.Fluent.IO;
@@ -18,17 +19,20 @@ using SysExtensions.Text;
 
 namespace Mutuo.Etl.Blob {
   public class AzureBlobFileStore : ISimpleFileStore {
-    public AzureBlobFileStore(Uri sas, StringPath pathSansContainer = null) 
-      : this(pathSansContainer) => Container = new CloudBlobContainer(sas);
+    public ILogger Log { get; }
 
-    public AzureBlobFileStore(string cs, StringPath path) : this(path) {
+    public AzureBlobFileStore(Uri sas, ILogger log, StringPath pathSansContainer = null) 
+      : this(pathSansContainer, log) => Container = new CloudBlobContainer(sas);
+
+    public AzureBlobFileStore(string cs, StringPath path, ILogger log) : this(path, log) {
       var storage = CloudStorageAccount.Parse(cs);
       var client = new CloudBlobClient(storage.BlobEndpoint, storage.Credentials);
       var containerName = path.Tokens.FirstOrDefault() ?? throw new InvalidOperationException("path needs to at least have a container");
       Container = client.GetContainerReference(containerName);
     }
 
-    AzureBlobFileStore(StringPath path) {
+    AzureBlobFileStore(StringPath path, ILogger log) {
+      Log = log;
       H = new HttpClient {
         Timeout = 10.Minutes()
       };
@@ -58,70 +62,22 @@ namespace Mutuo.Etl.Blob {
       }
     }
 
-    public async Task<T> Get<T>(StringPath path, bool zip = true) where T : class {
-      var fullPath = path.AddJsonExtention(zip);
-      var blob = BlobRef(fullPath);
-
-      await using var mem = new MemoryStream();
-      try {
-        await blob.DownloadToStreamAsync(mem);
-      }
-      catch (Exception) {
-        var exists = await blob.ExistsAsync();
-        if (!exists) return null;
-        throw;
-      }
-      mem.Position = 0;
-
-      Stream stream = mem;
-      if (zip) {
-        await using var zr = new GZipStream(mem, CompressionMode.Decompress);
-        stream = zr;
-      }
-      using var tr = new StreamReader(stream, Encoding.UTF8);
-      return JsonExtensions.DefaultSerializer.Deserialize<T>(tr);
-    }
 
     public CloudBlockBlob BlobRef(StringPath path) => Container.GetBlockBlobReference(BasePathSansContainer.Add(path));
-
-    /// <summary>Serializes item into the object store</summary>
-    /// <param name="path">The path to the object (no extensions)</param>
-    /// <param name="item"></param>
-    /// <param name="zip"></param>
-    public async Task Set<T>(StringPath path, T item, bool zip) {
-      await using var memStream = new MemoryStream();
-      
-      if (zip) {
-        await using (var zipWriter = new GZipStream(memStream, CompressionLevel.Optimal, leaveOpen:true)) {
-          await using var tw = new StreamWriter(zipWriter, Encoding.UTF8);
-          JsonExtensions.DefaultSerializer.Serialize(new JsonTextWriter(tw), item);
-        }
-      }
-      else {
-        await using (var tw = new StreamWriter(memStream, Encoding.UTF8, leaveOpen: true)) {
-          JsonExtensions.DefaultSerializer.Serialize(new JsonTextWriter(tw), item);
-        }
-      }
-      
-      var fullPath = path.AddJsonExtention(zip);
-      memStream.Seek(0, SeekOrigin.Begin);
-      var blob = BlobRef(fullPath);
-      AutoPopulateProps(fullPath, blob);
-      await blob.UploadFromStreamAsync(memStream);
-      
-      
-    }
+    
 
     public async Task Save(StringPath path, FPath file) {
       var blob = BlobRef(path);
       AutoPopulateProps(path, blob);
       await blob.UploadFromFileAsync(file.FullPath);
+      Log.Debug("Saved {Path}", path);
     }
 
     public async Task Save(StringPath path, Stream contents) {
       var blob = BlobRef(path);
       AutoPopulateProps(path, blob);
       await blob.UploadFromStreamAsync(contents);
+      Log.Debug("Saved {Path}", path);
     }
 
     public async Task<Stream> OpenForWrite(StringPath path) {
